pyproject.toml
[project]
name = "cyberbrain-mcp"
version = "1.1.0"
description = "Knowledge capture and retrieval for Claude sessions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Dan Singer"}
]
keywords = ["claude", "mcp", "knowledge-management", "obsidian"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastmcp>=3.0.0",
    "mcp",
    "pyyaml>=6.0",
    "ruamel.yaml",
]

[project.optional-dependencies]
bedrock = ["anthropic>=0.40.0"]
semantic = ["fastembed>=0.3", "usearch==2.23.0"]

[project.scripts]
cyberbrain-mcp = "cyberbrain.mcp.server:main"
cyberbrain-extract = "cyberbrain.extractors.extract_beats:main"
cyberbrain-reindex = "cyberbrain.extractors.search_index:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cyberbrain"]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
addopts = "--tb=no -q --no-header"
markers = [
    "core: Essential tests - always run in targeted mode",
    "extended: Integration tests - run on full regression only",
    "slow: Performance tests - run manually only",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.9",
    "basedpyright>=1.20",
    "pre-commit>=4.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "PTH"]
ignore = [
    "B008",   # function calls in default args — FastMCP Depends() pattern
    "E501",   # line length — handled by formatter for code, not for strings/comments
    "E402",   # module-level imports not at top — needed for sys.path manipulation in tests
    "B904",   # raise-from in except — implicit chaining used intentionally
    "B905",   # zip-strict — not applicable in all contexts
    "E741",   # ambiguous variable name — short names in comprehensions
    "F841",   # unused local variable — some assigned for readability
    "SIM102", # collapsible if — nested ifs often clearer
    "SIM103", # return bool directly — explicit if/else often clearer
    "SIM105", # use contextlib.suppress — try/except pass is explicit
    "SIM108", # ternary — if/else often clearer
    "SIM109", # compare-or — multiple equality comparisons sometimes clearer
    "B007",   # unused loop variable — rename to _var is cosmetic
    "SIM117", # nested with — separate with statements often clearer
    "PTH",    # pathlib replacements for os.path/open — too many changes for a formatting pass
    "UP038",  # isinstance union — fix requires --unsafe-fixes for some cases
    "UP042",  # StrEnum — fix requires --unsafe-fixes
]

[tool.ruff.lint.per-file-ignores]
"src/cyberbrain/extractors/extract_beats.py" = ["F401"]  # re-export hub
"src/cyberbrain/mcp/shared.py" = ["F401"]  # re-exports from extract_beats

[tool.ruff.format]
quote-style = "double"

[tool.basedpyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"